refactor(app): extract CrossRepoState into app/cross_repo.rs
Struct and its methods (known_hosts, resolve_repo_config,
resolve_local_path) move verbatim; fully-qualified paths meant no
new imports were needed. Re-exported from app. Part of #270.
